moved styles to their own file
