+++ betterplotlib/__init__.py
@@ -1,2 +1,3 @@
 from .tools import *
 from .colors import *
+from .styles import *
